+++ R/EpiTracer-package.R
@@ -15,5 +15,6 @@ utils::globalVariables(c(
   "episomal", "has_excision_scar",
   "chr", "chrom1", "chrom2", "start1", "start2", "strand1", "strand2",
   "strands", "gene", "gene_name", "pos_label", "min_coord", "max_coord",
-  "Polyploidy", "y", "label", "pos", "gieStain", "color"
+  "Polyploidy", "y", "label", "pos", "gieStain", "color",
+  "x", "xend", "yend", "lwd", "colour", "grp"
 ))